% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{codeblock}
\alias{codeblock}
\title{codeblock outlines a standardized code block}
\usage{
codeblock()
}
\value{
sends some code to the console to be copied to the source
}
\description{
codeblock outlines a standardized code block
}
\examples{
codeblock()
}
