% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{commoncode}
\alias{commoncode}
\title{commoncode generates initial code chunk at start of each chapter}
\usage{
commoncode()
}
\value{
nothing but it writes R syntax to the console ready to be copied to
the source window
}
\description{
commoncode solves the problem of having the same R environment
operating in each chapter. It generates an initial code chunk to be
placd at the start of each chapter.
}
\examples{
commoncode()
}
