% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{crossref}
\alias{crossref}
\title{crossref prints some help on cross referencing to the console}
\usage{
crossref()
}
\value{
nothing it just prints text to the console
}
\description{
crossref prints some help on cross referencing to the console
}
\examples{
crossref()
}
