% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{figureimport}
\alias{figureimport}
\title{figureimport generates Quarto code to import a figure from a file}
\usage{
figureimport()
}
\value{
nothing but it writes R syntax to the console ready to be copied to
the source window
}
\description{
figureimport generates Quarto to import a figure from a file
}
\examples{
figureimport()
}
