% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{figuresetup}
\alias{figuresetup}
\title{figuresetup generates Quarto syntax to generate a new figure}
\usage{
figuresetup()
}
\value{
nothing but it writes R syntax to the console ready to be copied to
the source window
}
\description{
figuresetup generates Quarto syntax to generate a new table
}
\examples{
 figuresetup()
}
