% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{getoutline}
\alias{getoutline}
\title{getoutline pulls out the headings from a Quarto file}
\usage{
getoutline(infile, rundir = "", outfile = "")
}
\arguments{
\item{infile}{the name of the Quarto file, this can be just the filename
with the qmd filetype or the complete path.}

\item{rundir}{if only the filename is given then the directory in which it
is to be found should be placed in 'rundir', default = ''}

\item{outfile}{if output to the console is wanted leave outfile as the
default = '', otherwise provide a filename as either a filename or the
complete path, depending on rundir.}
}
\value{
write the documents headings to the console ora file
}
\description{
getoutline generates a rough outline of a document by
identifying each line that begins with a heading, that is a #, ##, ###,
of more, and prints them out to the console of to a file.
}
\examples{
# syntax getoutline("stockassess_aphorisms.qmd",rundir="C:/")
# will send an outline of a file in the C:/ drive to the console.
}
