% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{importfigure}
\alias{importfigure}
\title{importfigure generates Quarto code to import a figure from a file}
\usage{
importfigure()
}
\value{
nothing but it writes R syntax to the console ready to be copied to
the source window
}
\description{
generates Quarto to import a figure from a file
}
\examples{
importfigure()
}
