% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{makeGitBook}
\alias{makeGitBook}
\title{makeGitBook generates a template set of files to make a GitBook}
\usage{
makeGitBook(
  rundir,
  filename = "manuscript.qmd",
  author = "Malcolm Haddon",
  affiliation = "IMAS, University of Tasmania",
  department = "IMAS-FA",
  city = "Hobart",
  DOCX = TRUE,
  HTML = TRUE,
  docref = "wordreferencestyle.docx"
)
}
\arguments{
\item{rundir}{the directory in which the quarto files are to be held. It
might also hold the wordreference file.}

\item{filename}{what name for the qmd file, default='manuscript.qmd'}

\item{author}{who is the first author, default = 'Malcolm Haddon'}

\item{affiliation}{the institutional affiliation, default =
'IMAS, University of Tasmania'}

\item{department}{the department within teh institution, default = IMAS-FA}

\item{city}{the city for the Institution, default = Hobart}

\item{DOCX}{should the option of a docx be included, default = TRUE}

\item{HTML}{should the option of an html document be included,
default = TRUE. If TRUE then a custom.scss file is also generated into
the rundir}

\item{docref}{the word document which is used as a reference if a docx
document is to be generated. It is recommended that this be placed
within the rundir. If this is not present in the rundir then the full
path and filename is required.}
}
\value{
nothing but it does generate four files into rundir
}
\description{
makeGitBook simplifies the creation of a GitBook. Such a book
is hosted in GitHub io pages and provides a hyperlinked document visible
on the web, but also provides to option of downloading a PDF of the
whole book. A GitBook is generated automatically from a GitHub
repository if that repository is appropriately structured. This function
aims to simplify the construction of the file structure necessary. More
details of how to instigate the production of R packages and GitHub
repositories will be provided in teh GitBook in which this function is
being described
}
\examples{
# syntax
# require(codeutils)
# require(qmdutils)
# usedir <- tempdir()
# makeGitBook(usedir)
# dir(usedir)
}
