% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{makeQuarto}
\alias{makeQuarto}
\title{makeQuarto generates a template for a quarto manuscript}
\usage{
makeQuarto(
  rundir,
  filename = "manuscript.qmd",
  title = "A Title",
  author = "Malcolm Haddon",
  affiliation = "IMAS, University of Tasmania",
  department = "IMAS-FA",
  city = "Hobart",
  DOCX = TRUE,
  PDF = TRUE,
  HTML = TRUE,
  docref = "wordreferencestyle.docx"
)
}
\arguments{
\item{rundir}{the directory in which the quarto files are to be held. It
might also hold the wordreference file.}

\item{filename}{what name for the qmd file, default='manuscript.qmd'}

\item{title}{what title for the manuscript, default = 'A Title'}

\item{author}{who is the first author, default = 'Malcolm Haddon'}

\item{affiliation}{the institutional affiliation, default =
'IMAS, University of Tasmania'}

\item{department}{the department within teh institution, default = IMAS-FA}

\item{city}{the city for the Institution, default = Hobart}

\item{DOCX}{should the option of a docx be included, default = TRUE}

\item{PDF}{should the option of a pdf be included, default = TRUE}

\item{HTML}{should the option of an html document be included,
default = TRUE. If TRUE then a custom.scss file is also generated into
the rundir}

\item{docref}{the word document which is used as a reference if a docx
document is to be generated. It is recommended that this be placed
within the rundir. If this is not present in the rundir then the full
path and filename is required.}
}
\value{
nothing but it does generate one or two files into rundir
}
\description{
makeQuarto generates a template for a quarto based manuscript
that can be any or all of html, pdf, of docx. If using docx it is
recommended that the word macros found at
https//:github.com/haddonm/quarto_macros, although these are still being
developed.
}
\examples{
require(codeutils)
usedir <- tempdir()
makeQuarto(usedir,title="An Ideal World")
txt <- readLines(pathtopath(usedir,"manuscript.qmd"))
nline <- length(txt)
for (i in 1:nline) print(txt[i],quote=FALSE)
}
