% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmdutils.R
\docType{package}
\name{rmdutils}
\alias{qmdutils}
\alias{qmdutils-package}
\title{qmdutils: a set of functions to assist with Quarto writing}
\description{
The qmdutils package provides an array of utility functions,
these include template functions that aid getting the syntax correct for
generating figures and tables within Quarto
}
\section{functions}{

\describe{
\item{figuresetup}{template for creating a figure}
\item{importfigure}{template for importing a figure file}
\item{tablesetup}{template for generating a table}
}
}

\author{
\strong{Maintainer}: Malcolm Haddon \email{malcolm.haddon@gmail.com} (\href{https://orcid.org/0000-0001-6971-7585}{ORCID})

}
\keyword{internal}
