% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{simpleblock}
\alias{simpleblock}
\title{simpleblock outlines a standardized code block}
\usage{
simpleblock()
}
\value{
sends some code to the console to be copied to the source
}
\description{
simpleblock outlines a standardized code block
}
\examples{
simpleblock()
}
