% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmd-utils.R
\name{tablesetup}
\alias{tablesetup}
\title{tablesetup outlines a standardized table block}
\usage{
tablesetup()
}
\value{
sends some code to the console to be copied to the source
}
\description{
tablesetup outlines a standardized table block
}
\examples{
tablesetup()
}
